Rename folder & add new IGame
