+++ src/package/implementation/PlayerImpl.ts
@@ -1,0 +1,5 @@
+import { IPlayer } from "../interface/IPlayer";
+
+class PlayerImpl implements IPlayer {
+
+}
