--- src/package/implementations/PlayerImpl.ts
@@ -1,5 +1,0 @@
-import { IPlayer } from "../interfaces/IPlayer";
-
-class PlayerImpl implements IPlayer {
-
-}
