+++ src/package/interface/IEventCenter.ts
