+++ src/package/interface/IGame.ts
@@ -1,0 +1,8 @@
+import { IRoom } from "./IRoom";
+
+export interface IGame {
+    createRoom(): boolean;
+
+    getRoom(id: number): IRoom;
+
+}
