+++ src/package/interface/IPlayer.ts
