+++ src/package/interface/IRoom.ts
